requirements.txt
numpy==1.18.5
tensorflow-gpu==1.15.4
tqdm==4.62.3
scikit-learn==1.0
networkx==2.6.3
matplotlib==3.4.3
seaborn==0.11.2
scipy==1.4.1
protobuf==3.20.0